rewrite the nora node engine service
